add pluralize for ingredients add functionality to merge ingredients

+++ app/controllers/tags_controller.rb
@@ -1,14 +1,60 @@
 class TagsController < ApplicationController
 
   include ApplicationHelper
+  include TagsHelper
 
   before_action :is_admin?, only: [:index, :destroy]
 
   PAGINATION = 25
 
+  def new
+    @tag = Tag.new(category: params[:category])
+  end
+
+  def create
+    c = params[:tag][:category] || 0
+    if params[:commit]
+      @tag = Tag.new(tag_params)
+      unless @tag.save
+        render :new
+        return
+      end
+    end
+    redirect_to path_from_category(c)
+  end
+
+  def edit
+    @tag = Tag.find(params[:id])
+    render :new
+  end
+
+  def update
+    c = params[:tag][:category] || 0
+    if params[:commit]
+      @tag = Tag.find(params[:id])
+      unless @tag.update(tag_params)
+        render :new
+        return
+      end
+    end
+    redirect_to path_from_category(c)
+  end
+
+  def merge
+    m = params[:merge].map { |e| e[0] }
+    if m && m.count > 1
+      Tagging.where(tag_id: m[1..-1]).update_all(tag_id: m[0])
+      Tag.where(id: m[1..-1]).destroy_all
+    end
+    redirect_to ingredients_path(merged: m[0])
+  end
+
   def index
+    if params[:merged]
+      @merged = params[:merged].to_i
+    end
     @category = params[:default][:category] || 0
-    @sort_by, @sort_direction = sort_by_column(tag: params[:tag], used: params[:used])
+    @sort_by, @sort_direction = sort_by_column(tag: params[:tag], other: params[:other], used: params[:used])
     if @sort_by == :used
       if @sort_direction == 1
         t =
@@ -29,16 +75,25 @@ class TagsController < ApplicationController
     @tags = t[@page*PAGINATION..@page*PAGINATION+PAGINATION-1]
   end
 
-  def delete
+  def destroy
     if params[:id].present?
       r = Tag.find(params[:id])
       if r
-        r.destroy
+        if r.taggings.count == 0
+          r.destroy
+        else
+          flash[:error] = t('views.tags.invalid_tag_has_recipes')
+        end
       else
         flash[:error] = t('view.tags.invalid_tag_id')
       end
     end
     redirect_to tags_path
-
   end
+
+  private
+
+    def tag_params
+      params.require(:tag).permit(:tag, :other, :category)
+    end
 end
